+++ app/studio/beta-34/page.tsx
@@ -1,0 +1,7 @@
+"use client";
+
+import Beta34Dashboard from "@/components/beta34/Beta34Dashboard";
+
+export default function Beta34Page() {
+  return <Beta34Dashboard />;
+}